Replace lodash-humps with more size-efficient implementation
The `lodash-humps` imports lodash function in a way not compatible with how we
do things in Calypso, i.e., from individual modules like `lodash/isArray`. That
results in a lot of code duplication in the base bundle.

This patch replaces `lodash-humps` with our own implementation (very inspired by
`lodash-humps`) that does the imports correctly.