Post Editor Visibility: remove the showingAcceptDialog instance property
The property is being set, but never read from. Looking at the initial OSS commit, the
`EditorVisibility` UI used to be a modal popover (it's a popup menu today) and the property
was used to cancel hiding the popover when an accept modal dialog is shown. Today, the UI
is different and this property is not needed.